chnaged directory structure of project so that it becomes pip installable
